backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
boto3==1.34.25
Pillow==10.2.0
qrcode[pil]==7.4.2
python-multipart==0.0.6
openai==1.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiofiles==23.2.1